api_gateway/app/middleware/correlation_id.py
# app/middleware/correlation_id.py
from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
import uuid


class CorrelationIdMiddleware(BaseHTTPMiddleware):
    """
    Middleware to add correlation ID to each request
    
    Correlation IDs help trace requests through distributed systems
    """
    
    async def dispatch(self, request: Request, call_next):
        # Get correlation ID from header or generate new one
        correlation_id = request.headers.get(
            "X-Correlation-ID",
            str(uuid.uuid4())
        )
        
        # Store in request state
        request.state.correlation_id = correlation_id
        
        # Process request
        response = await call_next(request)
        
        # Add correlation ID to response headers
        response.headers["X-Correlation-ID"] = correlation_id
        
        return response
