api_gateway/requirements_txt.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Message Queue
aio-pika==9.3.1

# Cache & Database
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Authentication
PyJWT==2.8.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
